some stable stuff

+++ exam-management-system/exam-management-system/src/main/java/com/exam/repository/SuperintendentRepository.java
@@ -4,6 +4,6 @@ import com.exam.model.SuperintendentModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface SuperintendentRepository extends JpaRepository<SuperintendentModel, Integer> {
-    Optional<SuperintendentModel> findByIdAndPassword(Integer id, String password);
+public interface SuperintendentRepository extends JpaRepository<SuperintendentModel, Long> {
+    Optional<SuperintendentModel> findByIdAndPassword(Long id, String password);
 }
