+++ exam-management-system/exam-management-system/src/main/java/com/exam/repository/TeacherRepository.java
@@ -4,6 +4,6 @@ import com.exam.model.TeacherModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface TeacherRepository extends JpaRepository<TeacherModel, Integer> {
-    Optional<TeacherModel> findByIdAndPassword(Integer id, String password);
+public interface TeacherRepository extends JpaRepository<TeacherModel, Long> {
+    Optional<TeacherModel> findByIdAndPassword(Long id, String password);
 }
